+++ src/main/java/API/com/example/E_COMMERCY/model/Cart.java
@@ -4,11 +4,8 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 @Table(name = "cart")
@@ -37,7 +34,7 @@ public class Cart {
     }
 
 
-    public int getId() {
+    public long getId() {
         return id;
     }
